+++ database/migrations/2017_11_13_170401_create_project_table.php
@@ -19,7 +19,7 @@ class CreateProjectTable extends Migration
                 ->comment('项目名称');
             $table->unsignedInteger('type_id')
                 ->comment('类型id');
-            $table->unsignedInteger('money')
+            $table->float('money')
                 ->comment('项目价格');
             $table->tinyInteger('status')
                 ->comment('项目状态 1. 进行中 2. 还款中 3. 已完结');
@@ -29,9 +29,9 @@ class CreateProjectTable extends Migration
                 ->comment('项目除期外的天数');
             $table->string('issue_explain')
                 ->comment('期说明');
-            $table->unsignedInteger('rental_money')
+            $table->float('rental_money')
                 ->comment('租房价格');
-            $table->unsignedInteger('collect_money')
+            $table->float('collect_money')
                 ->comment('收房价格');
             $table->integer('weight')->default(0)
                 ->comment('权重');
@@ -82,9 +82,9 @@ class CreateProjectTable extends Migration
             $table->increments('id');
             $table->string('name', 100)->unique()
                 ->comment('类型名称');
-            $table->unsignedInteger('max_money')
+            $table->float('max_money')
                 ->comment('最大金额');
-            $table->unsignedInteger('min_money')
+            $table->float('min_money')
                 ->comment('最小金额');
             $table->tinyInteger('repayment_type')
                 ->comment('还款方式 1. 等额本息 2. 先息后本');
@@ -130,11 +130,11 @@ class CreateProjectTable extends Migration
                 ->comment('用户id');
             $table->tinyInteger('issue_num')
                 ->comment('期数');
-            $table->unsignedInteger('money')
+            $table->float('money')
                 ->comment('金额');
-            $table->unsignedInteger('principal')
+            $table->float('principal')
                 ->comment('本金');
-            $table->unsignedInteger('interest')
+            $table->float('interest')
                 ->comment('利息');
             $table->tinyInteger('is_repayment')->default(1)
                 ->comment('是否还款 1. 未还款 2. 已还款');
